+++ src/components/shared/TimerSelect.vue
@@ -16,6 +16,7 @@
   const t = useTranslation()
   const timerSettings = [
     { label: 'None', value: null },
+    { label: '1s', value: 1 },
     { label: '5s', value: 5 },
     { label: '15s', value: 15 },
     { label: '30s', value: 30 },